requirements.txt
# Core ML Libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=2.0.0,<3.0.0
scikit-learn>=0.24.0
torchmetrics>=0.8.0  # PyTorch metrics collection for ML

# Configuration Management
hydra-core>=1.1.0
omegaconf>=2.1.0
# pyyaml>=5.4.0


# Foundation Image Model Access
timm>=0.9.9

# Data Processing
pandas>=1.3.0
pillow>=8.2.0
matplotlib>=3.4.0
scipy>=1.7.0
# seaborn>=0.11.0

# Experiment Tracking
wandb>=0.12.0
tqdm>=4.61.0

# Utilities
jsonschema>=3.2.0
requests>=2.25.0
colorama >= 0.4.6
einops >= 0.8.1
xformers>= 0.0.18
# typing-extensions>=3.10.0
# setuptools >= 75.9.1

# Jupyter Notebook Support
ipykernel>=6.0.0  # IPython kernel for Jupyter
ipython>=7.31.1   # Enhanced interactive Python shell
jupyter-client>=7.0.0  # Client library for Jupyter protocol
jupyter-core>=4.9.0    # Core functionality of Jupyter
matplotlib-inline>=0.1.3  # Matplotlib inline plotting support
jupyter>=1.0.0
ipywidgets>=7.6.0

# For loading environment variables from .env file
python-dotenv==1.0.0

# Development Tools (optional)
# Uncomment these if needed for development
# black==23.1.0
# isort==5.12.0
# flake8==6.0.0
# mypy==1.0.1
# pytest==7.2.2
# pytest-cov==4.0.0